retrieve stories from db and provide via ctx
Added loading state to useGame() so the game container can make decisions about what to render. When the the loading state received from useStories() evals to false, the "storiesLoaded" action updates status to "idle".

stories arr stored in useGame state. gameCount added, which can be used to access the correct story and will be incremented on action of type "win".